New architecture, now reload finally works (I hope at least)

+++ commands/shutdown.js
@@ -2,26 +2,22 @@
 const settings = require('../settings.json');
 const chalk = require('chalk');
 
-module.exports = {
-    Shutdown: async (message, client) => {
-        console.log('Fuckoff trigger');
-        if (settings.mods.includes(message.author.id)) {
-            await message.channel.send("Bye ! ");
-            console.log('disconnecting...');
-            await client.destroy();
-            console.log(chalk.red('Disconnected'));
-            await process.exit();
-        }
-        else {
-            await message.channel.send("ask for an admin role fam");
-            console.log(chalk.red(message.author.username + 'tried to shut me down'));
-        }
-    },
-    ShutdownTerminal: async (message, client)=> {
-      console.log('disconnecting...');
+exports.run = async (message,client,args,terminal) => {
+{
+    if(terminal){
       await client.destroy();
-      console.log(chalk.red('Disconnected'));
+      await console.log(chalk.cyan('Disconnected'));
       process.exit();
-
     }
+    else if (!terminal && settings.mods.includes(message.author.id)) {
+      await message.channel.send("Bye ! ");
+      await client.destroy();
+      console.log(chalk.cyan('Disconnected'));
+      await process.exit();
+    }
+    else {
+      await message.channel.send("ask for an admin role fam");
+      console.log(chalk.bgYellow(message.author.username + 'tried to shut me down'));
+    }
+
 };
